Move docker state tests to pytest
